tests(validators): added booking provider test
