+++ tests/unit-tests/UnitTests/Validators/IntegrityChecks/Form/GeneratedIdConfigurationTests.cs
@@ -3,7 +3,7 @@ using form_builder.Models;
 using form_builder.Validators.IntegrityChecks.Form;
 using Xunit;
 
-namespace form_builder_tests.UnitTests.Validators.IntegrityChecks
+namespace form_builder_tests.UnitTests.Validators.IntegrityChecks.Form
 {
     public class GeneratedIdConfigurationTests
     {
