+++ tests/unit-tests/UnitTests/Validators/IntegrityChecks/Form/HasDuplicateQuestionIdsCheckTests.cs
@@ -5,7 +5,7 @@ using form_builder.Validators.IntegrityChecks.Form;
 using form_builder_tests.Builders;
 using Xunit;
 
-namespace form_builder_tests.UnitTests.Validators.IntegrityChecks
+namespace form_builder_tests.UnitTests.Validators.IntegrityChecks.Form
 {
     public class HasDuplicateQuestionIdsCheckTests
     {
