+++ tests/unit-tests/UnitTests/Validators/IntegrityChecks/Form/PaymemtConfigurationCheckTests.cs
@@ -14,7 +14,7 @@ using form_builder_tests.Builders;
 using Moq;
 using Xunit;
 
-namespace form_builder_tests.UnitTests.Validators.IntegrityChecks
+namespace form_builder_tests.UnitTests.Validators.IntegrityChecks.Form
 {
     public class PaymemtConfigurationCheckTests
     {
