+++ tests/unit-tests/UnitTests/Validators/IntegrityChecks/Form/RetrieveExternalActionsCheckTests.cs
@@ -6,7 +6,7 @@ using form_builder_tests.Builders;
 using Moq;
 using Xunit;
 
-namespace form_builder_tests.UnitTests.Validators.IntegrityChecks
+namespace form_builder_tests.UnitTests.Validators.IntegrityChecks.Form
 {
     public class RetrieveExternalActionsCheckTests
     {
